% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDX-confidence-envelop.R
\name{exceedance_confidence}
\alias{exceedance_confidence}
\title{Computing the confidence envolop of the false discover proportion for the data}
\usage{
exceedance_confidence(profiled_data, alpha, ri = NULL, sri = NULL, rx = NULL)
}
\arguments{
\item{profiled_data}{an exceedance_profile object}

\item{alpha}{numeric, the confidence level}

\item{ri}{integer, the index of the rejected hypotheses, see details.}

\item{sri}{integer, the index of the ascending ordered p-values which the
corresponding hypotheses are rejected, see details.}

\item{rx}{numeric, the value of the pvalues which the
corresponding hypotheses are rejected, see details.}
}
\value{
a \code{1 - alpha} level confidence envolop
}
\description{
Computing the 1 - alpha level confidence envolop of the false discover
proportion(FDP) given a set of rejected hypotheses.
The confidence envolop can be viewed as a measurement of the quality of
the statistical inference.
}
\details{
This function is for constructing the confidence envolop of the
FDP given the set of rejected hypothese. The confidence envolop
depends on three factors:
\itemize{
\item The p-value samples
\item The confidence level \code{alpha}
\item The rejected hypotheses.
}
Therefore, given the data, confidence level and the hypotheses that you want
to reject, we can obtain a \code{1 - alpha} confidence envolop of the FDP.

The rejected hypotheses can be expressed in three ways. You can use the
original index \code{ri} to indicate which hypotheses you want to reject. For
example, if \code{ri = 1:2}, it means the first and second hypotheses are rejected.

However, in practice, it is more common to reject the hypotheses which
have small pvalues. You can achieve it by providing the parameter \code{sri}.
For example, if \code{sri = 1:2}, it means the hypothese which have the smallest
or second smallest pvalues are rejected. Alternatively, \code{rx} can be used if
you want to match the pvalues not the index. That is, a hypotheis is
rejected if its pvalue matches any value in \code{rx}.
}
\examples{
## The 3rd pvalue statistic
param <- param_fast_GW(statistic = "kth_p", param1 = 3)

## generate p-values
x <- rbeta(10, 1, 10)

## profile the data
profile <- profile_pvalue(x,param)


## compute the 95\% confidence envolop
alpha <- 0.05

## reject the first three hypotheses
exceedance_confidence(profile, alpha, ri = 3)

## reject the hypothese which pvalues are equal to
## the first three samples.
## In other word, this is equivalent to reject the first three hypotheses
exceedance_confidence(profile, alpha, rx = x[1:3])

## reject the hypotheses which have the lowest 3 p-values
exceedance_confidence(profile, alpha, sri = 3)


## Determine which hypotheses can be rejected while controlling the
## exceedance rate: P(FDP > bound) < alpha
alpha <- 0.05
bound <- 0.2
exceedance_inference(profile, alpha, bound)

}
