% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDX-inference.R
\name{exceedance_inference}
\alias{exceedance_inference}
\title{Perform multiple hypothesis testing while controlling the exceedance rate}
\usage{
exceedance_inference(profiled_data, alpha, bound)
}
\arguments{
\item{profiled_data}{an exceedance_profile object}

\item{alpha}{numeric, the confidence level}

\item{bound}{The upper bound of the false discover proportion}
}
\value{
Indices of the hypotheses that are rejected in the procedure.
}
\description{
Perform multiple hypothesis testing while controlling the exceedance rate.
The exceedance rate is Prob(FDP > bound), where FDP is the false discover
proportion defined by the number of false positives devided by the number
of total rejections. Small FDP with large number of reject is favorable
in practice.
}
\examples{
## The 3rd pvalue statistic
param <- param_fast_GW(statistic = "kth_p", param1 = 3)

## generate p-values
x <- rbeta(10, 1, 10)

## profile the data
profile <- profile_pvalue(x,param)


## compute the 95\% confidence envolop
alpha <- 0.05

## reject the first three hypotheses
exceedance_confidence(profile, alpha, ri = 3)

## reject the hypothese which pvalues are equal to
## the first three samples.
## In other word, this is equivalent to reject the first three hypotheses
exceedance_confidence(profile, alpha, rx = x[1:3])

## reject the hypotheses which have the lowest 3 p-values
exceedance_confidence(profile, alpha, sri = 3)


## Determine which hypotheses can be rejected while controlling the
## exceedance rate: P(FDP > bound) < alpha
alpha <- 0.05
bound <- 0.2
exceedance_inference(profile, alpha, bound)

}
